backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pyjwt==2.10.1
bcrypt==4.2.1
python-multipart==0.0.19
cryptography==44.0.0
apscheduler==3.10.4
aiofiles==24.1.0
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
